Integrate with udev via libudev++ (incomplete)
* configure.ac (KCIO_MODULES), (KC_KEYBOARD_MODULES): Add libudev++1.
* Makefile.am (kc_keyboard_kc_keyboard_SOURCES): Add devicepool.{cc,h}.
* kc-keyboard/devicepool.{cc,h}: New files, implementing the beginnings
of a class representing a dynamic pool of serial port devices available
on the system.
* kc-keyboard/inputwindow.cc (InputWindow::init_ui_actions): Extend the
status icon menu with a "Disconnected" item and a placeholder item for
a dynamic list of available serial ports.
* kc-keyboard/main.cc (main): Create Udev::Client and KC::DevicePool
objects and have the device pool receive uevents. Enumerate existing
devices on startup to populate the pool.