Check curses library for wide character support
* build/dk-curses.m4 (DK_MODULE_CURSES): Split up the Autoconf macro
into a main macro and a couple of helper macros.  Interpret an option
list parameter to allow influencing the behavior of the macro.  Provide
the require-wchar option to enable testing for wide character support.
Implement the check for wide character support as a link-time test.
* build/dk-feature.m4: New file, defining the feature test helper
macros from danielk's Autostuff.
* configure.ac (DK_MODULE_CURSES): Use the new require-wchar option
argument.