Send custom receive code to KC for V.24 boost mode
* libkc/v24boost.asm: New Z80 assembly language file with the code of
a custom V.24 receive program operating at 19200 Baud.  A simple means
of integrity checking is also provided.
* cli/kcsend.c: Implement the V.24 boost transfer mode by writing the
code of the custom receive program to the cassette buffer of the target
device, and then execute the program to receive the actual payload and
write it to memory.  Use boost mode by default, but provide a command
line option for turning it off.  Also allow autostart to be inhibited
via another command line flag.