Add a load offset command line parameter to kcsend
* cli/kcsend.c (parse_int16): New utility function for parsing a 16 bit
integer from a string.
(send_kcfile): Take a new "loadoffset" parameter and add it to the load
and end addresses read from the KCC file, and also to the start address
if the KCC file indicates that it should be translated.
(main): Parse the new "-o" command line option and its argument.