Support TAP and SSS tape file formats
* libkc/kctape.c: New file, defining helper routines shared by kcplay
and kcrec.
* cli/kcplay.c, cli/kcrec.c: Provide new command-line option -t FORMAT
to manually select a file format.  Rework the playback/recording code
to flexibly handle any of TAP, KCC or SSS.  Actually implement the SSS
format correctly -- it was there but broken before.  Rename write_*()
to play_*() and read_*() to record_*() for clarity.
* cli/kcrec.c: Fine-tune the FM decoding parameters to gracefully deal
with aberrations in the signal output of the original hardware.
* libkc/cmdline.c (kc_parse_arg_format): New shared helper function.
* libkc/libkc.h (KCFileFormat): New enumeration identifying a KC tape
file format and subformat.